+++ test/modelValidatorTests.ts
@@ -768,12 +768,21 @@ describe("Model Validation", () => {
     });
   });
 
-  describe("parameter required validation", () => {
+  describe("parameter validation", () => {
     it("should fail when missing write required parameters in example", async () => {
-      const specPath2 = `${testPath}/modelValidation/swaggers/specification/parameterRequiredValidation/test.json`;
-      const result = await validate.validateExamples(specPath2, undefined);
+      const specPath2 = `${testPath}/modelValidation/swaggers/specification/parameterValidation/test.json`;
+      const result = await validate.validateExamples(specPath2, "SqlServers_Get");
       assert.strictEqual(result.length, 1);
       assert.strictEqual(result[0].code, "REQUIRED_PARAMETER_EXAMPLE_NOT_FOUND");
+    });
+
+    it("should report error when operation has no own parameters", async () => {
+      const specPath2 = `${testPath}/modelValidation/swaggers/specification/parameterValidation/test.json`;
+      const result = await validate.validateExamples(specPath2, "SqlServers_Get_noOwnParameter");
+      assert.strictEqual(result.length, 1);
+      assert.strictEqual(result[0].code, "MAX_LENGTH");
+      assert.strictEqual(result[0].message, "String is too long (8 chars), maximum 4");
+      assert.strictEqual(result[0].exampleJsonPath, "$parameters.shortSubscriptionId");
     });
   });